First round of webpack and modularizing all of the files and classes
