+++ src/animations.js
@@ -1,0 +1,127 @@
+export class Animation {
+    constructor() { }
+
+    update(timeNow, timeElapsed) { }
+
+    spawnRenderables() { }
+
+    isExpired() { return true; }
+}
+
+export class TextAnimaton extends Animation {
+    constructor(centerX, centerY, text) {
+        super();
+        this.centerX = centerX;
+        this.centerY = centerY;
+        this.frameSpeed = 600;
+        this.lastFrame = Date.now();
+        this.isVisible = false;
+        this.text = text;
+    }
+
+    update(timeNow, timeElapsed) {
+        if (timeNow - this.lastFrame > this.frameSpeed) {
+            this.isVisible = !this.isVisible;
+            this.lastFrame = timeNow;
+        }
+    }
+
+    spawnRenderables() {
+        if (this.isVisible) {
+            let characters = [];
+            // add WIN in the center of the explosion.
+            for (var i = 0; i < this.text.length; i++) {
+                var x_offset = Math.floor(this.text.length / 2) - i;
+                characters.push(new StaticCharacter(this.centerX - x_offset, this.centerY, this.text.charAt(i)));
+            }
+            return characters;
+        }
+        return null;
+    }
+
+    isExpired() {
+        // this animation never goes away
+        return false;
+    }
+}
+
+export class WinAnimation extends Animation {
+    constructor(centerX, centerY, maxX, maxY) {
+        super();
+        this.radius = -1;
+        this.centerX = centerX;
+        this.centerY = centerY;
+        this.frameSpeed = 60;
+        this.lastFrame = Date.now();
+        this.maxX = maxX;
+        this.maxY = maxY;
+    }
+
+    update(timeNow, timeElapsed) {
+        var elapsed = timeNow - this.lastFrame;
+        if (elapsed >= this.frameSpeed) {
+            this.radius++;
+            this.lastFrame = timeNow;
+        }
+    }
+    
+    spawnRenderables() {
+        let characters = [];
+        // create explosion particles in a blast radius away from the center
+        for (var y = this.centerY - this.radius; y <= this.centerY + this.radius; y++) {
+            var difference = Math.abs(this.centerY - y);
+            var numXs = Math.min(2, this.radius - difference + 1); // add 1 because we always want at least 1 explosion on each line
+            
+            for (var i = 0; i < numXs; i++) {
+                var multiplier = 1;
+                if (i == 0) {
+                    multiplier = -1;
+                }
+                
+                var x = this.centerX + ((this.radius - difference) * multiplier);
+                var character = new StaticCharacter(x, y, '*');
+                characters.push(character);
+            }
+        }
+        return characters;
+    }
+
+    isExpired() {
+        return this.centerX - (this.radius * 2) < 0 &&
+            this.centerX + (this.radius * 2) > this.maxX &&
+            this.centerY - (this.radius * 2) < 0 &&
+            this.centerY + (this.radius * 2) > this.maxY;
+    }
+}
+
+export class AnimationHandler {
+    constructor(renderer) {
+        this.renderer = renderer;
+        this.animations = [];
+    }
+
+    addAnimation(animation) {
+        this.animations.push(animation);
+    }
+
+    update(timeNow, timeElapsed, gameObjects) {
+        for (var i = this.animations.length - 1; i >= 0 ; i--) {
+            var animation = this.animations[i];
+            animation.update(timeNow, timeElapsed);
+
+            if (animation.isExpired()) {
+                // remove it from our list
+                this.animations.splice(i, 1);
+            } else {
+                characters = this.animations[i].spawnRenderables();
+                if (characters !== null) {
+                    characters.map(x => gameObjects.addObject(x));
+                }
+            }
+        }
+    }
+
+    clearAnimations() {
+        this.animations = []
+    }
+}
