+++ src/facing.js
@@ -1,0 +1,4 @@
+export const FACING_UP = 'U';
+export const FACING_LEFT = 'L';
+export const FACING_DOWN = 'D';
+export const FACING_RIGHT = 'R';
