+++ src/game.js
@@ -1,0 +1,93 @@
+import {Thread} from "./thread.js";
+import {Renderer} from "./renderer.js";
+import {Rectangle} from "./rectangle.js";
+
+export class Game {
+    constructor() {
+        this.lastkeyPresses = [];
+        this.threadUpdate = null;
+        this.threadDraw = null;
+        this.renderer = null;
+    }
+
+    initialize(updateFunction, drawFunction, options) {
+        this.renderer = new Renderer(options['viewportWidth'], options['viewportHeight']);
+        this.threadUpdate = new Thread(updateFunction);
+        this.threadDraw = new Thread(drawFunction);
+
+        this.threadUpdate.start(options['updateFPS']); //update X times per second
+        this.threadDraw.start(options['drawFPS']); //draw X times per second
+
+        var that = this;
+
+        if (options['playInBrowser']) {
+            function keyDownHandler(e) {
+                var key = e.keyCode;
+                if (key) {
+                    that.lastkeyPresses.push(key);
+                }
+            };
+
+            // listen to the browser keys instad of direct console input
+            document.addEventListener("keydown", keyDownHandler, false);    
+        } else {
+            // this allows us to read keys directly from console input without ENTER
+            process.stdin.setRawMode(true); 
+            process.stdin.on('readable', function(data) {   
+                var key = process.stdin.read();
+                if (key) {
+                    that.lastkeyPresses.push(key);
+                }
+            });
+        }
+    }
+
+    getLastKeypress() {
+        if (this.lastkeyPresses.length <= 0) {
+            return null;
+        }
+        return this.lastkeyPresses.shift();
+    }
+
+    handleMovement(obj, intendedPosition, gameObjects) {
+        let isObstructed = false;
+        if (obj.obeysPhysics) {
+            const newRect = new Rectangle(intendedPosition.x, intendedPosition.y, obj.getBounds().width, obj.getBounds().height);
+            isObstructed = gameObjects.objects.filter(c => 
+                                                    c !== obj && 
+                                                    c.isPhysical && 
+                                                    c.getBounds().intersects(newRect)).length > 0;
+        }
+
+        if (!isObstructed) {
+            obj.getBounds().x = intendedPosition.x;
+            obj.getBounds().y = intendedPosition.y;
+        }
+
+        obj.intendedPosition = null;
+    }
+
+    update(now, timeElapsed, gameObjects) {
+        // update everything
+        gameObjects.objects.map(x => x.update(now, timeElapsed));
+
+        // handle movement intentions
+        gameObjects.objects.filter(x => x.intendedPosition != null).map(x => this.handleMovement(x, x.intendedPosition, gameObjects));
+
+        // check all collisions
+        gameObjects.objects.filter(x => x.hasOwnProperty('collider')).map(x => x.collider.checkCollision(gameObjects));
+
+        // remove everything that needs to be removed
+        var removableObjects = gameObjects.objects.filter(x => x.removeFromGameObjects);
+        if (removableObjects.length > 0) {
+            removableObjects.map(x => gameObjects.removeObject(x));
+            this.renderer.setIsDirty();
+        }
+
+        this.renderer.checkRedraw(gameObjects);
+    }
+
+    bogus() {
+        
+    }
+}
