+++ src/game_objects.js
@@ -1,0 +1,23 @@
+export class GameObjects {
+    constructor() {
+        this.objects = [];
+    }
+
+    addObject(obj) {
+        this.objects.push(obj);
+        for (var i = 0; obj.children && i < obj.children.length; i++) {
+            this.addObject(obj.children[i]);
+        }
+    }
+
+    removeObject(obj) {
+        this.objects = this.objects.filter(x => x !== obj);
+        for (var i = 0; obj.children && i < obj.children.length; i++) {
+            this.removeObject(obj.children[i]);
+        }
+    }
+
+    removeAllObjects() {
+        this.objects = [];
+    }
+}
