+++ src/map.js
@@ -1,0 +1,52 @@
+import {WallCharacter} from "./wall_character.js";
+
+export class Map {
+    constructor(mapData) {
+        this.mapData = mapData;
+        this.width = mapData.width;
+        this.height = mapData.height;
+        this.characters = null;
+    }
+
+    getMapCharacters() {
+        if (this.characters) {
+            return this.characters;
+        }
+
+        this.characters = [];
+
+        // add all of the objects within the level
+        for (var i = 0; i < this.mapData.map_objects.length; i++) {
+            var obj = this.mapData.map_objects[i];
+            for (var row = 0; row < obj.image.length; row++) {
+                var y = row + obj.y;
+                var rowStr = obj.image[row];
+                for (var col = 0; col < rowStr.length; col++) {
+                    var x = obj.x + col;
+                    var thisChar = rowStr.charAt(col);
+                    if (thisChar != ' ') {
+                        this.characters.push(new WallCharacter(x, y, thisChar));
+                    }
+                }
+            }
+        }
+
+        // add in left/right level barriers
+        for (var y = 0; y < this.height; y++) {
+            this.characters.push(new WallCharacter(0, y, "|"));
+            this.characters.push(new WallCharacter(this.width - 1, y, "|"));
+        }
+
+        // add in top/bottom level barriers
+        for (var x = 0; x < this.width; x++) {
+            this.characters.push(new WallCharacter(x, 0, "-"));
+            this.characters.push(new WallCharacter(x, this.height - 1, "-"));
+        }
+
+        return this.characters;
+    }
+
+    getIsWall(x, y) {
+        return this.getMapCharacters().filter(ch => ch.getBounds().intersectsPoint(x, y)).length > 0;
+    }
+}
