+++ src/math_extensions.js
@@ -1,0 +1,3 @@
+export function randomNumber(max) {
+    return Math.floor((Math.random() * max) + 1);
+}
