+++ src/projectile_character.js
@@ -1,0 +1,50 @@
+import {Character} from "./character.js";
+import {Collider} from "./collider.js";
+import {Movable} from "./movable.js";
+import {Sprite} from "./sprite.js";
+
+export class ProjectileCharacter extends Character {
+    constructor(initialX, initialY, direction, maxDistance, spriteMap) {
+        super(initialX, initialY);
+        
+        this.isVisible = true;
+        this.direction = direction;
+        this.distanceTraveled = 0;
+        this.maxDistance = maxDistance;
+        this.travelSpeed = (1 / 6.0) * 1000;
+        this.travelCounter = 0;
+        this.obeysPhysics = false;
+
+        // set up our sprite
+        this.sprite = new Sprite(spriteMap, this);
+        this.sprite.setState(direction);
+        this.children.push(this.sprite);
+        
+        // set up our ability to move
+        this.movable = new Movable(this);
+        this.children.push(this.movable);
+
+        // check for collisions with objects
+        // TODO: add this to children?
+        this.collider = new Collider(this);
+    }
+
+    think() {
+        this.movable.move(this.direction);
+        this.distanceTraveled++;
+        if (this.distanceTraveled >= this.maxDistance) {
+            this.removeFromGameObjects = true;
+        }
+    }
+    
+    update(timeNow, timeElapsed) {
+        super.update(timeNow, timeElapsed);
+
+        this.travelCounter = this.travelCounter + timeElapsed;
+
+        if (this.travelCounter >= this.travelSpeed) {
+            this.travelCounter = this.travelCounter - this.travelSpeed;
+            this.think();
+        }
+    }
+}
