+++ src/rectangle.js
@@ -1,0 +1,28 @@
+export class Rectangle {
+    constructor(x, y, width, height) {
+        this.x = x;
+        this.y = y;
+        this.width = width;
+        this.height = height;
+    }
+
+    intersectsPoint(x, y) {
+        if (x < this.x + this.width &&
+            x >= this.x &&
+            y < this.y + this.height &&
+            y >= this.y) {
+            return true;
+        }
+        return false;
+    }
+
+    intersects(rectangle) {
+        if (this.x + this.width <= rectangle.x ||
+            this.x >= rectangle.x + rectangle.width ||
+            this.y + this.height <= rectangle.y ||
+            this.y >= rectangle.y + rectangle.height) {
+            return false;
+        }
+        return true;
+    }
+}
