+++ src/renderer.js
@@ -1,0 +1,65 @@
+import {Rectangle} from "./rectangle.js";
+
+export class Renderer {
+    constructor(viewW, viewH) {
+        this.viewport = new Rectangle(0, 0, viewW, viewH);
+        this.isDirty = true;
+    }
+
+    clearScreen() {
+        // clear the screen and set cursor at 0,0
+        console.clear();
+    }
+
+    isOnScreen(character) {
+        return character.isVisible && character.getBounds().intersects(this.viewport);
+    }
+
+    render(gameObjects) {
+        var output = '';
+        var renderableObjects = this.getRenderableObjectsOnScreen(gameObjects);
+        for (var row = this.viewport.y; row < this.viewport.y + this.viewport.height; row++) {
+            for (var col = this.viewport.x; col < this.viewport.x + this.viewport.width; col++) {
+                var characters = renderableObjects.filter(c => c.getCharacter).map(c => c.getCharacter(row, col)).filter(c => c != null);
+                if (characters.length > 0) {
+                    output = output + characters[0];
+                } else {
+                    output = output + ' ';
+                }
+            }
+            output = output + '\n';
+        }
+
+        console.log(output);
+
+        // clear the redraw flag on all objects we were able to render
+        renderableObjects.map(c => c.needsRedraw = false);
+    }
+
+    getIsDirty() {
+        return this.isDirty;
+        
+    }
+
+    setIsDirty() {
+        this.isDirty = true;
+    }
+
+    checkRedraw(gameObjects) {
+        if (this.getRenderableObjectsOnScreen(gameObjects).filter(c => c.needsRedraw).length > 0) {
+            this.setIsDirty();
+        }
+    }
+
+    getRenderableObjectsOnScreen(gameObjects) {
+        var that = this;
+        var renderable = gameObjects.objects.filter(c => c.isVisible && 
+                                                    that.isOnScreen(c));
+        return renderable;
+    }
+
+    centerViewportOn(character, map) {
+        this.viewport.x = Math.min(map.width - this.viewport.width, Math.max(0, character.getX() - (this.viewport.width / 2)));
+        this.viewport.y = Math.min(map.height - this.viewport.height, Math.max(0, character.getY() - (this.viewport.height / 2)));
+    }
+}
