+++ src/sprite.js
@@ -1,0 +1,137 @@
+export class Sprite {
+    // format of a spriteMap:
+    // {
+    //     "<state>": [{ "displayTime": <timeInMillisec>, "characters": [<row1string>, <row2string>, etc] },
+    //                 { "displayTime": <timeInMillisec>, "characters": [<row1string>, <row2string>, etc] } ]
+    // }
+    constructor(spriteMap, parentObject) {
+        this.spriteMap = spriteMap;
+        this.parentObject = parentObject;
+        this.state = 0;
+        this.stateElapsed = 0;
+        this.frame = 0;
+        this.isVisible = true;
+    }
+
+    getBounds() {
+        return this.parentObject.bounds;
+    }
+
+    update(timeNow, timeElapsed) {
+        this.stateElapsed = this.stateElapsed + timeElapsed;
+        
+        var prevFrame = this.frame;
+        this.frame = this.calculateCurrentFrame();
+        if (this.frame != prevFrame) {
+            this.parentObject.onAnimated();
+        }
+
+        // TODO: figure out a better way to do colliding with sprites
+        //this.calculateSize();
+    }
+
+    calculateSize() {
+        var sprites = this.spriteMap.states[this.state];
+        var frame = this.frame;
+        var characterRows = sprites[frame]["characters"];
+
+        var firstRow = 999999;
+        var lastRow = 0;
+        var firstCol = 999999;
+        var lastCol = 0;
+
+        for (var row = 0; row < characterRows.length; row++) {
+            for (var col = 0; col < characterRows[row].length; col++) {
+                var symbol = characterRows[row].charAt(col);
+                if (symbol && symbol != ' ') {
+                    firstRow = Math.min(firstRow, row);
+                    lastRow = Math.max(lastRow, row);
+                    firstCol = Math.min(firstCol, col);
+                    lastCol = Math.max(lastCol, col);
+                }
+            }
+        }
+
+        // TODO: this feels dirty
+        this.parentObject.getBounds().width = lastCol - firstCol + 1;
+        this.parentObject.getBounds().height = lastRow - firstRow + 1;
+    }
+
+    setState(newState) {
+        this.state = newState;
+    }
+
+    getX() {
+        return this.parentObject.getX();
+    }
+
+    getY() {
+        return this.parentObject.getY();
+    }
+
+    calculateCurrentFrame() {
+        var sprites = this.spriteMap.states[this.state];
+        var totalTime = sprites.reduce(function(acc, curVal) { return acc + curVal["displayTime"]; }, 0);
+        var leftover = this.stateElapsed % totalTime;
+        var frame = 0;
+        var timeAccumulator = 0;
+        for (var i = 0; i < sprites.length; i++) {
+            frame = i;
+            // TODO: replace this with a reduce()
+            var displayTime = sprites[i]["displayTime"];
+            if (timeAccumulator + displayTime > leftover) {
+                break;
+            }
+            timeAccumulator = timeAccumulator + displayTime;
+        }
+
+        return frame;
+    }
+
+    getAnchoredX() {
+        var characterRows = this.spriteMap.states[this.state][this.frame]["characters"];
+
+        // Remove any notion of the map location and localize to this sprite only
+        var ourCol = this.parentObject.getX();
+        if (this.spriteMap.anchor == "center") {
+            // assumes the first row is the same width as the other frames
+            ourCol = ourCol - Math.floor(characterRows[0].length / 2);
+        }
+
+        return ourCol;
+    }
+
+    getAnchoredY() {
+        var characterRows = this.spriteMap.states[this.state][this.frame]["characters"];
+
+        // Remove any notion of the map location and localize to this sprite only
+        var ourRow = this.parentObject.getY();
+        if (this.spriteMap.anchor == "center") {
+            ourRow = ourRow - Math.floor(characterRows.length / 2);
+        }
+
+        return ourRow;
+    }
+
+    getCharacter(row, col) {
+        var sprites = this.spriteMap.states[this.state];
+        var frame = this.frame;
+        var characterRows = sprites[frame]["characters"];
+
+        // Remove any notion of the map location and localize to this sprite only
+        var ourRow = row - this.getAnchoredY();
+        var ourCol = col - this.getAnchoredX();
+
+        if (ourRow >= 0 && 
+            ourCol >= 0 && 
+            ourRow < characterRows.length &&
+            ourCol < characterRows[ourRow].length) {
+                var symbol = characterRows[ourRow].charAt(ourCol);
+                if (symbol != ' ') {
+                    return symbol;
+                }
+        }
+
+        return null;
+    }
+}
