+++ src/static_character.js
@@ -1,0 +1,15 @@
+import {Character} from "./character.js";
+
+export class StaticCharacter extends Character {
+    constructor(initialX, initialY, symbol) {
+         super(initialX, initialY);
+         this.symbol = symbol;
+     }
+ 
+     getCharacter(row, col) {
+         if (this.getX() == col && this.getY() == row) {
+             return this.symbol;
+         }
+         return null;
+     }
+ }
