+++ src/thread.js
@@ -1,0 +1,23 @@
+export class Thread {
+    constructor(functionPointer) {
+        this.functionPointer = functionPointer;
+    }
+
+    start(desiredFramerate) {
+        var that = this;
+        var minimumMillsecPerFrame = 1000/ desiredFramerate;
+
+        var internalRun = function() {
+            var now = Date.now();
+            that.functionPointer();
+            var after = Date.now();
+
+            // if the function call takes a while, reduce the delay until the next execute
+            var nextDelay = Math.max(0, minimumMillsecPerFrame - (after - now));
+            setTimeout(internalRun, nextDelay);
+        };
+
+        // initial call
+        setTimeout(internalRun, 0); 
+    }
+}
