+++ src/wall_character.js
@@ -1,0 +1,22 @@
+import {Character} from "./character.js";
+
+export class WallCharacter extends Character {
+    constructor(initialX, initialY, symbol) {
+        super(initialX, initialY);
+        this.symbol = symbol;
+        this.isPhysical = true;
+    }
+
+    getCharacter(row, col) {
+        if (this.getX() == col && this.getY() == row) {
+            return this.symbol;
+        }
+        return null;
+    }
+
+    collide(withObject) {
+        if (withObject instanceof ProjectileCharacter) {
+            withObject.removeFromGameObjects = true;
+        }
+    }
+}
