+++ webpack.config.js
@@ -1,0 +1,7 @@
+module.exports = {
+  entry: './src/treasure-hunt.js',
+  output: {
+    filename: 'bundle.js'
+  },
+  devtool: 'inline-source-map',
+};
